Remove `set` in favour of `Timeline` widget. Add track header support.
Removes the `set` function in favour of a `Timeline` widget builder that
allows for a more general approach to setting timeline parameters.

A `header` builder was added that allows for specifying a track header
area to the left of the tracks, along with a new `TracksCtx` that
assists with instantiation.

The `Timeline` trait was rename to `TimelineApi` pending a more suitable
name.